improve: type safety, move semantics, constructor correctness across codebase
- result.hpp: add explicit to all constructors, SFINAE or_else for void-returning
  functors, add move/copy constructors
- token_bucket.hpp: default burst=1.0 (was 0), remove unused <stdexcept>
- protocol.hpp: status_code unsigned int -> uint16_t
- tls_context.hpp: replace magic int min_tls_version with TlsVersion enum
- vault_client.hpp: timeout_seconds int -> uint32_t
- Add explicit move/delete to 8 classes: MetricsCollector, ShardedMetricsCollector,
  IoContextPool, ConnectionPool, Tracer, VaultClient, ScenarioEngine,
  PrometheusExporter, AuthProvider
- io_context_pool: guard get_context() against empty vector
- sharded_collector: fix non-standard zero-length array, latency_sum_us -> uint64_t
- http_client: add static_cast for uint16_t status_code narrowing

+++ core/metrics/sharded_collector.cpp
@@ -44,13 +44,12 @@ void ShardedMetricsCollector::record_request(uint16_t status_code,
     } else {
         auto cur_sum = shard.latency_sum_us.load(std::memory_order_relaxed);
         while (true) {
-            if (cur_sum > INT64_MAX - lat_count) {
-                // Would overflow — clamp to INT64_MAX
-                shard.latency_sum_us.store(INT64_MAX, std::memory_order_relaxed);
+            if (cur_sum > UINT64_MAX - static_cast<uint64_t>(lat_count)) {
+                shard.latency_sum_us.store(UINT64_MAX, std::memory_order_relaxed);
                 break;
             }
             if (shard.latency_sum_us.compare_exchange_weak(
-                    cur_sum, cur_sum + lat_count, std::memory_order_relaxed))
+                    cur_sum, cur_sum + static_cast<uint64_t>(lat_count), std::memory_order_relaxed))
                 break;
         }
     }
