+++ core/scenario/engine.cpp
@@ -108,7 +108,7 @@ public:
                             auto done = std::make_shared<std::atomic<bool>>(false);
                             client->async_request(*capture_req,
                                 [capture_req, &metrics, &callback, done, step](std::error_code ec, net::Response resp) mutable {
-                                    metrics.record_request(static_cast<uint16_t>(resp.status_code), resp.latency,
+                                    metrics.record_request(resp.status_code, resp.latency,
                                                            capture_req->body.size(), resp.body.size());
                                     if (callback) {
                                         callback(step, resp, metrics);
